Updated 3/25/18
Continued investigation into data. Trying to find the best way to filter out false tweets. Explored the number of people who posted a particular number of times. Thought about using TweetTokenizer to parse tweets to help with filtering. Remember that entries need to include the phrase "'week #" or "wk #" in their tweet to signal entry for a given week. Explored this with the post of Rodrigo Calloni